1. dealt with safety warning 2. Changed the data management of TMesh to MemoryPool 3. Fixed a bug in ++ operator of MPInterator 4. CMesh can read directorly from vertex and face list from memory 5. Fixed the initialization bug of PropPool, where it will overwrite all the new element with initial value. When prop is a class carrying in stack data, this can cause problem. 6. Added proppool to TMesh elements . Moved CPoint4 out of TMeshLib

+++ MeshFrame/MeshFrame/core/Mesh/Vertex.h
@@ -174,7 +174,7 @@ inline CHalfEdge *  CVertex::most_ccw_in_halfedge()
 	CHalfEdge * he = m_halfedge->ccw_rotate_about_target();
 	CHalfEdge * pStartHE = he;
 	//rotate to the most ccw in halfedge
-	int numHEs = m_outHEs.size();
+	size_t numHEs = m_outHEs.size();
 	int numHEIterated = 0;
 	while( he != NULL )
 	{
